refactor: modularize scripts and styles
