+++ index.html
@@ -21,7 +21,8 @@
     />
     <meta http-equiv="Pragma" content="no-cache" />
     <meta http-equiv="Expires" content="0" />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="styles/base.css" />
+    <link rel="stylesheet" href="styles/chat.css" />
     <script
       src="https://embed.twitch.tv/embed/v1.js"
       integrity="sha384-RZ28VvtH8iX008KoSoTGq9yUWWmruBCja+k14R2Q3EI55YSPQMv+lMsbVxJojPWh"
@@ -131,7 +132,7 @@
     </div>
     <div id="clickMe">click me</div>
     <canvas id="visualizer"></canvas>
-    <script src="script.js"></script>
+    <script type="module" src="src/main.js"></script>
     <div id="feedbackModal">
       <textarea
         id="feedbackInput"
