+++ styles/base.css
@@ -382,98 +382,3 @@ body {
   float: right;
 }
 
-#chat {
-  width: 420px;
-  height: 200px; /* initial size, can be resized by user */
-  overflow: auto;
-  border-radius: 8px;
-  background: rgba(0, 0, 0, 0.6);
-  color: #fff;
-  font-family: sans-serif;
-  padding: 8px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  position: relative;
-}
-
-#chatResizeHandle {
-  position: absolute;
-  top: 0;
-  right: 0;
-  width: 12px;
-  height: 12px;
-  cursor: nwse-resize;
-  background: rgba(255, 255, 255, 0.6);
-  border-bottom-left-radius: 4px;
-}
-#messages {
-  flex: 1;
-  overflow-y: auto;
-  margin-bottom: 8px;
-  font-size: 14px;
-  padding-right: 12px;
-  box-sizing: border-box;
-  scrollbar-gutter: stable;
-}
-#messages div {
-  overflow-wrap: anywhere;
-  word-break: break-word;
-}
-#chatForm {
-  display: flex;
-}
-#chatInput {
-  flex: 1;
-  padding: 4px;
-  font-size: 14px;
-}
-#chatForm button {
-  padding: 4px 8px;
-  margin-left: 4px;
-  font-size: 14px;
-}
-.chat-emote {
-  height: 1.5em;
-  vertical-align: middle;
-}
-.mention-highlight {
-  background: #ff0;
-  color: #000;
-  font-weight: bold;
-  padding: 0 2px;
-  border-radius: 2px;
-}
-#mentionSuggestions {
-  position: absolute;
-  bottom: 36px;
-  left: 8px;
-  right: 8px;
-  background: #222;
-  border: 1px solid #555;
-  display: none;
-  max-height: 100px;
-  overflow-y: auto;
-  z-index: 10;
-  font-size: 14px;
-}
-#mentionSuggestions div {
-  padding: 2px 4px;
-  cursor: pointer;
-}
-#mentionSuggestions div:hover {
-  background: #444;
-}
-#upgradePlaceholder {
-  display: flex;
-  align-items: center;
-  font-family: sans-serif;
-  font-weight: bold;
-  pointer-events: none;
-  margin-bottom: 8px;
-}
-#upgradePlaceholder img {
-  height: 20px;
-  width: 20px;
-  margin-right: 8px;
-}
